pyproject.toml
[project]
name = "customer-agent"
version = "1.0.0"
description = "电商AI客服"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.52.0",
    "websockets>=10.4",
    "requests>=2.28.0",
    "PySocks>=1.7.1",
    "cozepy>=0.15.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=5.0.1",
    "pyqt6>=6.9.0",
    "pyqt6-fluent-widgets[full]>=1.8.1",
]
